+++ uninstall.php
@@ -46,6 +46,8 @@ $npe_remove_site_data = static function () {
 	delete_option( 'npe_cache_warmup_queue_lock' );
 	delete_option( 'npe_cache_operations_state' );
 	delete_option( 'npe_performance_samples' );
+	delete_option( 'npe_dom_analysis_queue' );
+	delete_option( 'npe_dom_analysis_queue_lock' );
 
 	if ( defined( 'WP_CONTENT_DIR' ) ) {
 		// phpcs:disable WordPress.WP.AlternativeFunctions -- Uninstall must remove lock-protected generated files without loading plugin services.